Change fixed sized scrollback to command line flag
The fixed 4kB scrollback buffer size is now gone, the default will just
use 1kB of memory (which is still pretty decent). Most of the time
nowadays I am using tmux (or screen) in my terminal, so it does not matter
anymore how bug the buffer is. For those who would still want to be able
to have a bigger scrollback buffer, there is now a command line flag to
change the size.