Implement initial header bar support
As an option, uses GTK 3.8+ window header bar support to define a custom
title bar for dwt windows. For the moment, it only adds a close button and
the window title, using a custom widget. The idea is to be able to add
additional widgets, e.g. tint the title on beep or change the color of the
title bar, a tiny application menu as a gear icon, etc.