Ensure that new font size is initialized before use
In "font_size()", make sure that the font's "new_size" is always initialized
before using it. In the case where the font is reset to its initial size the
variable was not being initialized. This was mostly harmless, as re-applying
the window resize constraint metrics is a no-op when they are unchanged. The
fix makes the code properly correct, avoids making unneeded calls, and avoid
a compiler warning.