Fix window hang when exiting terminal via EOF
The signal signature for "child-exited" was changed in VTE 2.91 and it now
accepts an extra arguments with the exit status of the child process. Adding
the needed additional argument avoids the window not being closed due to
the fact that the exit status was being passed as a pointer to
gtk_widget_destroy().