Use DgSettings instead of DirDB for configuration
Using DgSettings avoids one extra layer as DgSettings uses properties
instead of custom API that calls into DirDB underneath. What is more,
because it uses properties, it is possible to connect signal handlers
to the "::notify" detail of the settings (which, remember, are properties)
and get notifications of the changes. Of course, it is also possible to
use bindings to maintain the values of properties synchronized with other
objects.