Avoid double-freeing a string in configure_term_widget()
The "opt_font" string is marked as scope-freeable memory using dg_lmem, so
if the pointer is not NULL, g_free() is always called on it. This means that
the format argument to g_variant_dict_lookup() cannot by "&s" to borrow
a reference to the string, because then g_free() will be called twice on the
string: once via "opt_font", and another when the dictionary is destroyed.
Of course, the double-free causes a segfault inside the libc.

To fix this, we always get a copy of the string using the plain "s" format
with g_variant_dict_lookup()