Support "profiles" for multiple configurations
This adds support for reading configuration files from an alternate
configuration directory. For that, argv[0] is inspected, and if it
is different from "dwt", a different D-Bus application ID is used
(org.perezdecastro.org.${argv[0]}) and the configuration directory
is changed to be $XDG_CONFIG_DIR/dwt.${argv[0]}. This allows to have
different sets of settings, and let dwt pick one according to the
name it has been invoked as.